변경 사항: `Sensor` 영역에 gatewayId를 String 타입에서 Long 타입으로 변환 과정 수행

+++ src/test/java/com/nhnacademy/sensor/repository/SensorRepositoryTest.java
@@ -84,7 +84,7 @@ class SensorRepositoryTest {
         Assertions.assertTrue(actual.isEmpty());
     }
 
-    private Sensor get(Integer sensorNo) {
+    private Sensor get(Long sensorNo) {
         Optional<Sensor> optional = sensorRepository.findById(sensorNo);
         Assertions.assertTrue(optional.isPresent());
         return optional.get();
